Move image function to image.py
