+++ pygmt/src/__init__.py
@@ -6,6 +6,7 @@ Source code for PyGMT modules.
 from pygmt.src.coast import coast
 from pygmt.src.colorbar import colorbar
 from pygmt.src.grdcontour import grdcontour
+from pygmt.src.image import image
 from pygmt.src.legend import legend
 from pygmt.src.logo import logo
